requirement.txt
aadict==0.2.3
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569679365/work
bcrypt @ file:///tmp/abs_6fpe92qzzo/croots/recipe/bcrypt_1659554336150/work
cffi @ file:///croot/cffi_1700254295673/work
click==8.1.7
cryptography @ file:///croot/cryptography_1694444244250/work
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.6
passlib==1.7.4
psycopg==3.1.14
pyasn1==0.5.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy-audit==0.1.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
python-multipart
